LayoutTests/fast/dom/timer-throttling-lowPowerMode-expected.txt

Tests that DOM timers are throttled in low power mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Enabling low power mode
internals.setLowPowerModeEnabled(true)
* Nesting: 1
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 2
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 3
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 4
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 5
PASS internals.isTimerThrottled(timerHandle) is true
* Nesting: 6
PASS internals.isTimerThrottled(timerHandle) is true

Disabling low power mode
internals.setLowPowerModeEnabled(false)
* Nesting: 7
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 8
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 9
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 10
PASS internals.isTimerThrottled(timerHandle) is false
PASS successfullyParsed is true

TEST COMPLETE
